requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
httpx>=0.28.1
pydantic>=2.11.0
python-dotenv>=1.0.0
loguru>=0.7.0
